Refactor: remaster layout for all pages
