+++ web/src/app/(loan-form)/contact-details/page.tsx
@@ -1,0 +1,8 @@
+"use client";
+import { NextPage } from "next";
+
+const ContactDetails: NextPage = () => {
+  return <div>Contact Details Form Lives Here</div>;
+};
+
+export default ContactDetails;
