+++ web/src/app/(static)/layout.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import "./globals.scss";
+import "../globals.scss";
 
 const inter = Inter({ subsets: ["latin"] });
