+++ web/src/app/components/FormNext/FormNext.tsx
@@ -1,8 +1,18 @@
+import { useRouter } from "next/navigation";
 import { useFormContext } from "react-hook-form";
-import { PrimaryButton } from "../shared/PrimaryButton";
 
-export const FormNext = () => {
+import { PrimaryButton } from "../shared/PrimaryButton";
+import { FormNextProps } from "./FormNext.types";
+
+export const FormNext = ({ nextLink }: FormNextProps) => {
+  const router = useRouter();
+
   const onSubmit = (data: any) => {
+    if (nextLink) {
+      router.push(nextLink);
+      return;
+    }
+
     console.log(data);
   };
