+++ web/src/app/components/FormNext/FormNext.types.ts
@@ -1,0 +1,3 @@
+export interface FormNextProps {
+  nextLink?: string;
+}
