--- web/src/app/layouts/LoanForm/index.tsx
@@ -1,1 +1,0 @@
-export { LoanForm } from "./LoanForm";
